refactor(lex-tables): decrease lex table state size
An earlier commit grew state_type temporarily to allow refactoring
handlers. The refactoring is done, so now shrink state_type back from 16
bits to 8 bits.

+++ src/quick-lint-js/fe/lex-tables.h
@@ -179,8 +179,7 @@ struct lex_tables {
   static_assert(character_class_table[static_cast<std::uint8_t>(u8'|')] == character_class::pipe);
   // clang-format on
 
-  // TODO(strager): Shrink to 8 bits.
-  using state_type = std::uint16_t;
+  using state_type = std::uint8_t;
 
   // How many bits in the state are reserved for the state number for
   // intermediate and non-unique terminal states, or for extra data for unique
